deepcorr/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pypi_0
astor=0.8.1=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2023.11.17=hbcca054_0
gast=0.2.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.8.0=pypi_0
importlib-metadata=6.7.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=13.2.0=h807b86a_3
libgomp=13.2.0=h807b86a_3
libstdcxx-ng=13.2.0=h7e041cc_3
libzlib=1.2.13=hd590300_5
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
ncurses=6.4=h59595ed_2
numpy=1.16.6=pypi_0
openssl=1.0.2u=h516909a_0
opt-einsum=3.3.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
protobuf=3.20.0=pypi_0
python=3.7.0=hd21baee_1006
readline=7.0=hf8c457e_1001
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
sqlite=3.28.0=h8b20d00_0
tensorboard=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
tensorflow-gpu=1.15.0=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tqdm=4.66.1=pypi_0
typing-extensions=4.7.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.16.0=pypi_0
xz=5.2.6=h166bdaf_0
zipp=3.15.0=pypi_0
zlib=1.2.13=hd590300_5
